Quota double-delete fix
BUG=142310


Review URL: https://chromiumcodereview.appspot.com/10832407

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@152532 0039d316-1c4b-4281-b951-d872f2087c98

+++ webkit/quota/quota_task.cc
@@ -31,7 +31,8 @@ void QuotaTask::Start() {
 
 QuotaTask::QuotaTask(QuotaTaskObserver* observer)
     : observer_(observer),
-      original_task_runner_(base::MessageLoopProxy::current()) {
+      original_task_runner_(base::MessageLoopProxy::current()),
+      delete_scheduled_(false) {
 }
 
 void QuotaTask::CallCompleted() {
@@ -49,6 +50,10 @@ void QuotaTask::Abort() {
 }
 
 void QuotaTask::DeleteSoon() {
+  DCHECK(original_task_runner_->BelongsToCurrentThread());
+  if (delete_scheduled_)
+    return;
+  delete_scheduled_ = true;
   MessageLoop::current()->DeleteSoon(FROM_HERE, this);
 }
